+++ code/test/sixsq/nuvla/server/resources/credential_infrastructure_service_swarm_lifecycle_test.clj
@@ -1,13 +1,14 @@
 (ns sixsq.nuvla.server.resources.credential-infrastructure-service-swarm-lifecycle-test
   (:require
     [clojure.data.json :as json]
-    [clojure.test :refer [deftest is use-fixtures]]
+    [clojure.test :refer [deftest is testing use-fixtures]]
     [peridot.core :refer [content-type header request session]]
     [sixsq.nuvla.server.app.params :as p]
     [sixsq.nuvla.server.middleware.authn-info :refer [authn-info-header]]
     [sixsq.nuvla.server.resources.credential :as credential]
     [sixsq.nuvla.server.resources.credential-template :as ct]
     [sixsq.nuvla.server.resources.credential-template-infrastructure-service-swarm :as cred-tpl]
+    [sixsq.nuvla.server.resources.job :as job]
     [sixsq.nuvla.server.resources.lifecycle-test-utils :as ltu]
     [sixsq.nuvla.server.util.metadata-test-utils :as mdtu]))
 
@@ -16,6 +17,7 @@
 
 
 (def base-uri (str p/service-context credential/resource-type))
+(def job-base-uri (str p/service-context job/resource-type))
 
 
 (deftest check-metadata
@@ -60,47 +62,47 @@
                                              :cert   cert-value
                                              :key    key-value}}]
 
-    ;; admin/user query should succeed but be empty (no credentials created yet)
-    (doseq [session [session-admin session-user]]
-      (-> session
+    (testing "Admin/user query should succeed but be empty (no credentials created yet)"
+      (doseq [session [session-admin session-user]]
+        (-> session
+            (request base-uri)
+            (ltu/body->edn)
+            (ltu/is-status 200)
+            (ltu/is-count zero?)
+            (ltu/is-operation-present :add)
+            (ltu/is-operation-absent :delete)
+            (ltu/is-operation-absent :edit))))
+
+    (testing "Anonymous credential collection query should not succeed"
+      (-> session-anon
           (request base-uri)
           (ltu/body->edn)
-          (ltu/is-status 200)
-          (ltu/is-count zero?)
-          (ltu/is-operation-present :add)
-          (ltu/is-operation-absent :delete)
-          (ltu/is-operation-absent :edit)))
+          (ltu/is-status 403)))
 
-    ;; anonymous credential collection query should not succeed
-    (-> session-anon
-        (request base-uri)
-        (ltu/body->edn)
-        (ltu/is-status 403))
+    (testing "Creating a new credential without reference will fail for all types of users"
+      (doseq [session [session-admin session-user session-anon]]
+        (-> session
+            (request base-uri
+                     :request-method :post
+                     :body (json/write-str create-import-no-href))
+            (ltu/body->edn)
+            (ltu/is-status 400))))
 
-    ;; creating a new credential without reference will fail for all types of users
-    (doseq [session [session-admin session-user session-anon]]
-      (-> session
+    (testing "Creating a new credential as anon will fail; expect 400 because href cannot be accessed"
+      (-> session-anon
           (request base-uri
                    :request-method :post
-                   :body (json/write-str create-import-no-href))
+                   :body (json/write-str create-import-href))
           (ltu/body->edn)
           (ltu/is-status 400)))
 
-    ;; creating a new credential as anon will fail; expect 400 because href cannot be accessed
-    (-> session-anon
-        (request base-uri
-                 :request-method :post
-                 :body (json/write-str create-import-href))
-        (ltu/body->edn)
-        (ltu/is-status 400))
-
-    ;; create a credential as a normal user
-    (let [resp    (-> session-user
-                      (request base-uri
-                               :request-method :post
-                               :body (json/write-str create-import-href))
-                      (ltu/body->edn)
-                      (ltu/is-status 201))
+    (let [resp    (testing "Create a credential as a normal user"
+                    (-> session-user
+                        (request base-uri
+                                 :request-method :post
+                                 :body (json/write-str create-import-href))
+                        (ltu/body->edn)
+                        (ltu/is-status 201)))
           id      (ltu/body-resource-id resp)
           uri     (-> resp
                       (ltu/location))
@@ -119,31 +121,37 @@
             (ltu/is-operation-present :edit)
             (ltu/is-operation-present :check)))
 
-      ;; ensure credential contains correct information
-      (let [credential (-> session-user
-                           (request abs-uri)
-                           (ltu/body->edn)
-                           (ltu/is-status 200))
-            {:keys [name description tags
-                    ca cert key
-                    parent]} (ltu/body credential)]
-        (is (= name name-attr))
-        (is (= description description-attr))
-        (is (= tags tags-attr))
-        (is (= ca ca-value))
-        (is (= cert cert-value))
-        (is (= key key-value))
-        (is (= parent parent-value))
+      (let [credential (testing "Ensure credential contains correct information"
+                         (-> session-user
+                             (request abs-uri)
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             (ltu/is-key-value :name name-attr)
+                             (ltu/is-key-value :description description-attr)
+                             (ltu/is-key-value :tags tags-attr)
+                             (ltu/is-key-value :ca ca-value)
+                             (ltu/is-key-value :cert cert-value)
+                             (ltu/is-key-value :key key-value)
+                             (ltu/is-key-value :parent parent-value)))]
 
-        ;; ensure that the check action works
-        (let [op-url    (ltu/get-op credential "check")
-              check-url (str p/service-context op-url)]
+        (testing "Ensure that the check action works"
+          (let [op-url    (ltu/get-op credential "check")
+                check-url (str p/service-context op-url)]
 
+            (-> session-user
+                (request check-url
+                         :request-method :post)
+                (ltu/body->edn)
+                (ltu/is-status 202)))))
+
+      (testing "Ensure that create and check credential created 2 cred check jobs"
+        (let [descr-changed "descr changed"
+              job-url-filter (str job-base-uri "?filter=action='credential_check'&target-resource/href='" id "'")]
           (-> session-user
-              (request check-url
-                       :request-method :post)
+              (request job-url-filter
+                       :request-method :put)
               (ltu/body->edn)
-              (ltu/is-status 202))))
+              (ltu/is-count 2))))
 
       ;; delete the credential
       (-> session-user